requirements.txt
sentence-transformers==2.2.2
langchain
langchain-pinecone
langchain_community
langchain_google_genai
langchain_huggingface
flask
pypdf
python-dotenv
streamlit

-e .